suscridash-backend/requirements.txt
flask
flask-cors
pyjwt
python-dotenv
werkzeug